refactored the narada and narada-pyodide packages so they share code
