+++ packages/narada-core/README.md
@@ -1,0 +1,3 @@
+# Narada Core
+
+Core models and actions for the Narada Python SDK.
