--- packages/narada-pyodide/src/narada/errors.py
@@ -1,6 +1,0 @@
-class NaradaError(Exception):
-    pass
-
-
-class NaradaTimeoutError(NaradaError):
-    pass
