+++ packages/narada/src/narada/__init__.py
@@ -1,6 +1,7 @@
 from narada.client import Narada
 from narada.config import BrowserConfig
-from narada.errors import (
+from narada_core.models import Agent
+from narada_core.errors import (
     NaradaError,
     NaradaExtensionMissingError,
     NaradaExtensionUnauthenticatedError,
@@ -8,12 +9,10 @@ from narada.errors import (
     NaradaTimeoutError,
     NaradaUnsupportedBrowserError,
 )
-from narada.models import Agent
+from narada_core.responses import Response, ResponseContent
 from narada.window import (
     LocalBrowserWindow,
     RemoteBrowserWindow,
-    Response,
-    ResponseContent,
 )
 
 __all__ = [
