+++ packages/narada/src/narada/window.py
@@ -9,7 +9,7 @@ import aiohttp
 from playwright.async_api import BrowserContext
 from pydantic import BaseModel
 
-from narada.actions.models import (
+from narada_core.actions.models import (
     AgenticSelectorAction,
     AgenticSelectorRequest,
     AgenticSelectors,
@@ -23,41 +23,27 @@ from narada.actions.models import (
     ReadGoogleSheetResponse,
     WriteGoogleSheetRequest,
 )
+from narada_core.models import Agent, RemoteDispatchChatHistoryItem, UserResourceCredentials
+from narada_core.errors import NaradaError, NaradaTimeoutError
+from narada_core.responses import (
+    Response,
+    ResponseContent,
+    Usage,
+    _StructuredOutput,
+    _MaybeStructuredOutput,
+    _ResponseModel,
+)
+from narada_core.window import BaseBrowserWindow
 from narada.config import BrowserConfig
-from narada.errors import NaradaError, NaradaTimeoutError
-from narada.models import Agent, RemoteDispatchChatHistoryItem, UserResourceCredentials
 
-_StructuredOutput = TypeVar("_StructuredOutput", bound=BaseModel)
-
-_MaybeStructuredOutput = TypeVar("_MaybeStructuredOutput", bound=BaseModel | None)
+# Note: Core types now imported from narada_core above
 
 
-class ResponseContent(TypedDict, Generic[_MaybeStructuredOutput]):
-    text: str
-    structuredOutput: _MaybeStructuredOutput
-
-
-class Usage(TypedDict):
-    actions: int
-    credits: int
-
-
-class Response(TypedDict, Generic[_MaybeStructuredOutput]):
-    requestId: str
-    status: Literal["success", "error", "input-required"]
-    response: ResponseContent[_MaybeStructuredOutput] | None
-    createdAt: str
-    completedAt: str | None
-    usage: Usage
-
-
-_ResponseModel = TypeVar("_ResponseModel", bound=BaseModel)
-
-
-class BaseBrowserWindow(ABC):
+class NaradaBrowserWindow(BaseBrowserWindow):
+    """Browser window implementation for the narada package using aiohttp."""
+    
     _api_key: str
     _base_url: str
-    _browser_window_id: str
 
     def __init__(
         self,
@@ -66,51 +52,9 @@ class BaseBrowserWindow(ABC):
         base_url: str,
         browser_window_id: str,
     ) -> None:
+        super().__init__(browser_window_id=browser_window_id)
         self._api_key = api_key
         self._base_url = base_url
-        self._browser_window_id = browser_window_id
-
-    @property
-    def browser_window_id(self) -> str:
-        return self._browser_window_id
-
-    @overload
-    async def dispatch_request(
-        self,
-        *,
-        prompt: str,
-        agent: Agent | str = Agent.OPERATOR,
-        clear_chat: bool | None = None,
-        generate_gif: bool | None = None,
-        output_schema: None = None,
-        previous_request_id: str | None = None,
-        chat_history: list[RemoteDispatchChatHistoryItem] | None = None,
-        additional_context: dict[str, str] | None = None,
-        time_zone: str = "America/Los_Angeles",
-        user_resource_credentials: UserResourceCredentials | None = None,
-        callback_url: str | None = None,
-        callback_secret: str | None = None,
-        timeout: int = 120,
-    ) -> Response[None]: ...
-
-    @overload
-    async def dispatch_request(
-        self,
-        *,
-        prompt: str,
-        agent: Agent | str = Agent.OPERATOR,
-        clear_chat: bool | None = None,
-        generate_gif: bool | None = None,
-        output_schema: type[_StructuredOutput],
-        previous_request_id: str | None = None,
-        chat_history: list[RemoteDispatchChatHistoryItem] | None = None,
-        additional_context: dict[str, str] | None = None,
-        time_zone: str = "America/Los_Angeles",
-        user_resource_credentials: UserResourceCredentials | None = None,
-        callback_url: str | None = None,
-        callback_secret: str | None = None,
-        timeout: int = 120,
-    ) -> Response[_StructuredOutput]: ...
 
     async def dispatch_request(
         self,
@@ -210,146 +154,6 @@ class BaseBrowserWindow(ABC):
         except asyncio.TimeoutError:
             raise NaradaTimeoutError
 
-    @overload
-    async def agent(
-        self,
-        *,
-        prompt: str,
-        agent: Agent | str = Agent.OPERATOR,
-        clear_chat: bool | None = None,
-        generate_gif: bool | None = None,
-        output_schema: None = None,
-        time_zone: str = "America/Los_Angeles",
-        timeout: int = 120,
-    ) -> AgentResponse[None]: ...
-
-    @overload
-    async def agent(
-        self,
-        *,
-        prompt: str,
-        agent: Agent | str = Agent.OPERATOR,
-        clear_chat: bool | None = None,
-        generate_gif: bool | None = None,
-        output_schema: type[_StructuredOutput],
-        time_zone: str = "America/Los_Angeles",
-        timeout: int = 120,
-    ) -> AgentResponse[_StructuredOutput]: ...
-
-    async def agent(
-        self,
-        *,
-        prompt: str,
-        agent: Agent | str = Agent.OPERATOR,
-        clear_chat: bool | None = None,
-        generate_gif: bool | None = None,
-        output_schema: type[BaseModel] | None = None,
-        time_zone: str = "America/Los_Angeles",
-        timeout: int = 120,
-    ) -> AgentResponse:
-        """Invokes an agent in the Narada extension side panel chat."""
-        remote_dispatch_response = await self.dispatch_request(
-            prompt=prompt,
-            agent=agent,
-            clear_chat=clear_chat,
-            generate_gif=generate_gif,
-            output_schema=output_schema,
-            time_zone=time_zone,
-            timeout=timeout,
-        )
-        response_content = remote_dispatch_response["response"]
-        assert response_content is not None
-
-        return AgentResponse(
-            status=remote_dispatch_response["status"],
-            text=response_content["text"],
-            structured_output=response_content.get("structuredOutput"),
-            usage=AgentUsage.model_validate(remote_dispatch_response["usage"]),
-        )
-
-    async def agentic_selector(
-        self,
-        *,
-        action: AgenticSelectorAction,
-        selectors: AgenticSelectors,
-        fallback_operator_query: str,
-        # Larger default timeout because Operator can take a bit to run.
-        timeout: int | None = 60,
-    ) -> None:
-        """Performs an action on an element specified by the given selectors, falling back to using
-        the Operator agent if the selectors fail to match a unique element.
-        """
-        return await self._run_extension_action(
-            AgenticSelectorRequest(
-                action=action,
-                selectors=selectors,
-                fallback_operator_query=fallback_operator_query,
-            ),
-            timeout=timeout,
-        )
-
-    async def go_to_url(
-        self, *, url: str, new_tab: bool = False, timeout: int | None = None
-    ) -> None:
-        """Navigates the active page in this window to the given URL."""
-        return await self._run_extension_action(
-            GoToUrlRequest(url=url, new_tab=new_tab), timeout=timeout
-        )
-
-    async def print_message(self, *, message: str, timeout: int | None = None) -> None:
-        """Prints a message in the Narada extension side panel chat."""
-        return await self._run_extension_action(
-            PrintMessageRequest(message=message), timeout=timeout
-        )
-
-    async def read_google_sheet(
-        self,
-        *,
-        spreadsheet_id: str,
-        range: str,
-        timeout: int | None = None,
-    ) -> ReadGoogleSheetResponse:
-        """Reads a range of cells from a Google Sheet."""
-        return await self._run_extension_action(
-            ReadGoogleSheetRequest(spreadsheet_id=spreadsheet_id, range=range),
-            ReadGoogleSheetResponse,
-            timeout=timeout,
-        )
-
-    async def write_google_sheet(
-        self,
-        *,
-        spreadsheet_id: str,
-        range: str,
-        values: list[list[str]],
-        timeout: int | None = None,
-    ) -> None:
-        """Writes a range of cells to a Google Sheet."""
-        return await self._run_extension_action(
-            WriteGoogleSheetRequest(
-                spreadsheet_id=spreadsheet_id, range=range, values=values
-            ),
-            timeout=timeout,
-        )
-
-    @overload
-    async def _run_extension_action(
-        self,
-        request: ExtensionActionRequest,
-        response_model: None = None,
-        *,
-        timeout: int | None = None,
-    ) -> None: ...
-
-    @overload
-    async def _run_extension_action(
-        self,
-        request: ExtensionActionRequest,
-        response_model: type[_ResponseModel],
-        *,
-        timeout: int | None = None,
-    ) -> _ResponseModel: ...
-
     async def _run_extension_action(
         self,
         request: ExtensionActionRequest,
@@ -389,7 +193,7 @@ class BaseBrowserWindow(ABC):
         return response_model.model_validate_json(response.data)
 
 
-class LocalBrowserWindow(BaseBrowserWindow):
+class LocalBrowserWindow(NaradaBrowserWindow):
     _config: BrowserConfig
     _context: BrowserContext
 
@@ -423,7 +227,7 @@ class LocalBrowserWindow(BaseBrowserWindow):
         await side_panel_page.reload()
 
 
-class RemoteBrowserWindow(BaseBrowserWindow):
+class RemoteBrowserWindow(NaradaBrowserWindow):
     def __init__(self, *, browser_window_id: str, api_key: str | None = None) -> None:
         super().__init__(
             api_key=api_key or os.environ["NARADA_API_KEY"],
